+++ projects/ui/src/lib/oui/scrollbar/scrollbar.ts
@@ -7,11 +7,15 @@ import {
 
 @Component({
   templateUrl: './scrollbar.html',
-  selector: 'oui-scrollbar',
+  selector: '[oui-scrollbar]',
   exportAs: 'OuiScrollbar',
   styleUrls: ['scrollbar.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
-  encapsulation: ViewEncapsulation.None
+  encapsulation: ViewEncapsulation.None,
+  host: {
+    class: 'oui-scrollbar-container',
+    '[style.height.px]': 'height'
+  }
 })
 export class OuiScrollbar {
   @Input() height: number;