+++ CHANGELOG.md
@@ -5,6 +5,116 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.10.3] - 2019-02-14
+
+### Added
+
+NA
+
+### Changed
+
+- NA
+
+### Removed
+
+- NA
+
+### Fixed
+
+- Fixed slide toggle color issue
+- Added FormsModule support in slide toggle
+- Dialog UI issue fixes [oncehub-5234](https://scheduleonce.atlassian.net/browse/ONCEHUB-5234)
+
+## [1.10.2] - 2019-02-12
+
+### Added
+
+NA
+
+### Changed
+
+- NA
+
+### Removed
+
+- NA
+
+### Fixed
+
+- Fixed slide toggle issue
+
+## [1.10.1] - 2019-02-08
+
+### Added
+
+NA
+
+### Changed
+
+- NA
+
+### Removed
+
+- NA
+
+### Fixed
+
+- Fixed overlay panel issue [oncehub-5476](https://scheduleonce.atlassian.net/browse/ONCEHUB-5476)
+
+## [1.10.0] - 2019-02-11
+
+### Added
+
+- [`oui-datepicker`](projects/ui/src/lib/oui/datepicker/README.md) (@garvit-rajput)
+
+### Changed
+
+- NA
+
+### Removed
+
+- NA
+
+### Fixed
+
+- NA
+
+## [1.9.1] - 2019-02-08
+
+### Added
+
+NA
+
+### Changed
+
+- NA
+
+### Removed
+
+- NA
+
+### Fixed
+
+- Fixed input field border color issue [oncehub-4400](https://scheduleonce.atlassian.net/browse/ONCEHUB-4400)
+
+## [1.9.0] - 2019-02-06
+
+### Added
+
+Table, Paginator, Sort directives and components
+
+### Changed
+
+- default dialog-scroll strategy
+
+### Removed
+
+- NA
+
+### Fixed
+
+- Fixed dialog scroll [ONCEHUB-5234](https://scheduleonce.atlassian.net/browse/ONCEHUB-5234)
+
 ## [1.8.6] - 2019-01-30
 
 ### Added